pyproject.toml
[tool.poetry]
name = "reserva-de-nome"
version = "0.0.1"
description = "Protótipo para reserva de nome de empresa (Moçambique)"
authors = ["World Bank Group <evicentegoncalve@worldbank.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.6"
dj_database_url = "^0.5.0"
django = "^3.1.1"
django-extensions = "^3.0.9"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
python-decouple = "^3.3"
whitenoise = {extras = ["brotoli"], version = "^5.2.0"}

[tool.poetry.dev-dependencies]
black = { version = "*", allow-prereleases = true }
createnv = "^0.0.1"
ipython = "^7.18.1"
mixer = "^6.1.3"
pdbpp = "^0.10.2"
pytest-black = "^0.3.11"
pytest-cov = "^2.10.1"
pytest-django = "^3.10.0"
pytest-flake8 = "^1.0.6"
pytest-watch = "^4.2.0"

[tool.pytest.ini_options]
flake8-ignore = ["E501" , "W503"]
DJANGO_SETTINGS_MODULE = "reserva.settings"
addopts = "--black --flake8 --cov reserva --no-cov-on-fail --cov-fail-under=80"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
